Move entity search surface tests into entities package (#294)
